chore(hid): inline LogReader into Supports.swift to avoid Xcode project changes
